feat: getServerToken
